requirements.txt
bcrypt==3.2.0
fastapi==0.109.2
httpx==0.23.1
lorem==0.1.1
Levenshtein==0.23.0
uvicorn==0.27.1
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pytest==7.2.0
pytest_asyncio==0.23.5
python-jose[cryptography]==3.3.0
python-multipart==0.0.7
